requirements.txt
Flask
openai
requests
beautifulsoup4
scikit-learn
python-dotenv